Remove undiscovered devices
Added functionality that removes devices from the display window when
they are no longer discoverable. However it still stores it in the
ArrayList DataDevices as DataRelation Objects. So we need to write the
data in the DataDevices ArrayList to a text file when we click the
expand view so it could be accessed by the other window